requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
distlib==0.3.9
Django==5.1.6
django-environ==0.12.0
django-filter==24.3
django-jazzmin==3.0.1
django-modeltranslation==0.19.12
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
filelock==3.17.0
flake8==7.1.1
identify==2.6.7
idna==3.10
inflection==0.5.1
mccabe==0.7.0
nodeenv==1.9.1
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pre_commit==4.1.0
psycopg2==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
sorl-thumbnail==12.11.0
sqlparse==0.5.3
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.29.2
